Module artemis::exchanges

This module contains the default exchanges Note that they require the default-exchanges feature

Structs

CacheExchange
The default caching exchange.

DedupExchange
The default deduplication exchange This will keep track of in-flight queries and catch any identical queries before they execute, instead waiting for the result from the in-flight query

FetchExchange
The default fetch exchange

TerminatorExchange
The terminating exchange. This will always be the last exchange in the chain and will simply return an error if called.

Traits

Client
Client trait passed to exchanges. Only exposes methods useful to exchanges