Trait artemis::types::ExchangeFactory

pub trait ExchangeFactory<TNext: Exchange> {
    type Output: Exchange;
    fn build(self, next: TNext) -> Self::Output;
}

An exchange factory. This must be passed to the ClientBuilder by the user, so it should take only necessary parameters.

In a scenario that doesn't require additional configuration this should just be a zero size struct

Associated Types

type Output: Exchange

Loading content...
Required methods

fn build(self, next: TNext) -> Self::Output

Loading content...