Crate artemis

A modern GraphQL Client with common built-in features as well as the ability to extend its functionality through exchanges

Build

This crate uses code generation to take your GraphQL files and turn them into strongly typed Rust modules. These contain the query struct, a zero-size type such as GetConference, as well as a submodule containing the Variables, any input types, the ResponseData type and any involved output types.

Having a strongly typed compile time representation with additional info (such as the __typename of all involved types and an abstract selection tree) means that the work the CPU has to do at runtime is very minimal, only amounting to serialization, deserialization and simple lookups using the statically generated data.

For details on how to use the query builder, see artemis-build

Exchanges

Exchanges are like a bi-directional middleware. They act on both the incoming and outgoing queries, passing them on if they can't return a result themselves.

There are three default exchanges, called in this order:

DedupExchange

The deduplication exchange (DedupExchange) filters out unnecessary queries by combining multiple identical queries into one. It does so by keeping track of in-flight queries and, instead of firing off another identical query, waiting for their results instead. This reduces network traffic, especially in larger applications where the same query may be used in multiple places and run multiple times simultaneously as a result.

CacheExchange

The cache exchange is a very basic, un-normalized cache which eagerly invalidates queries. It's focused on simplicity and correctness of data, so if a query uses any of the same types as a mutation it will always be invalidated by it. This means that especially if you have large amounts of different entities of the same type, this can become expensive quickly. For a more advanced normalized cache that invalidates only directly related entities see the artemis-normalized-cache crate.

FetchExchange

The fetch exchange will serialize the query, send it over the network and deserialize the response. This works on x86 using reqwest, or fetch if you're using WASM. This should be your last exchange in the chain, as it never forwards a query.

Re-exports

pub use client::Client;

pub use types::*;

Modules

client

exchanges
This module contains the default exchanges Note that they require the default-exchanges feature

types

Macros

ext
Creates a new ExtensionMap and fills it with the passed values

wasm_client

Structs

ClientBuilder
A builder for the artemis client

Error
An element in the top-level errors array of a response body.

Location
Represents a location inside a query string. Used in errors. See Error.

QueryBody
The form in which queries are sent over HTTP in most implementations. This will be built using the GraphQLQuery trait normally.

QueryError
A query error wrapper that allows for cheap and easy cloning across threads If a std::error::Error is needed, use QueryError.compat()

Response
The generic shape taken by the responses of GraphQL APIs.

Enums

PathFragment
Part of a path in a query. It can be an object key or an array index. See Error.

Traits

GraphQLQuery
A convenience trait that can be used to build a GraphQL request body. This will be implemented for you by codegen. It is implemented on the struct you place the derive on.

Functions

progressive_hash
When we have separate values it's useful to run a progressive version of djb2 where we pretend that we're still looping over the same value